Build using nix
Adds flake.nix and a Dockerfile using it
to either build locally using [nix](https://nixos.org)
or using docker, which essentially downloads
nixos and runs the same build process. The download
size is much smaller than the full texlive as
only needed packages are downloaded. They are defined
in nix/texlive.nix and should be enough to install
even outside of nix environment.

This also adds a devshell that you can enter using
`nix develop` to get an environment with texlive with
needed packages